examples/python/ingest_with_config/requirements.txt
python-dotenv==1.0.1
sift @ git+https://github.com/sift-stack/sift.git#subdirectory=python
